refactor: split DNA UI owner
