+++ scripts/quality-baseline.json
@@ -7,6 +7,6 @@
   "viewRuntimeBridgeLookups": 0,
   "labStateAppFiles": 0,
   "labStateTestFiles": 0,
-  "largeJsFilesOver800Lines": 14,
+  "largeJsFilesOver800Lines": 13,
   "maxJsFileLines": 1168
 }
